# server/openapi.yaml
openapi: 3.0.3
info:
  title: OpenTriologue API
  version: 1.0.0
  description: |
    OpenAPI contract for BYOA and project-scoped collaboration flows.
    Focus in this first version: health, BYOA message ingest, project access,
    team invite, and task create/update.
servers:
  - url: http://localhost:4001
    description: Local API
  - url: https://opentriologue.ai
    description: Production API
...
paths:
...
  /api/batch/messages:
    post:
      tags: [Batch]
      summary: Send messages to multiple rooms in one request
      operationId: postBatchMessages
      security:
        - BearerJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchMessageRequest"
      responses:
        "200":
          description: Batch processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchMessageResponse"
        "400":
          description: Invalid payload or too many messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/batch/agents/{mentionKey}/context:
    get:
      tags: [Batch]
      summary: Fetch consolidated execution context for one agent
      description: |
        Requires authentication. Caller must be the agent account behind
        `mentionKey` or an admin user. Returns rooms, tasks, and memory
        snippets relevant to the agent's active project scope.
      operationId: getAgentExecutionContext
      security:
        - BearerJwt: []
        - BearerByoa: []
      parameters:
        - $ref: "#/components/parameters/MentionKey"
      responses:
        "200":
          description: Agent context payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentContextResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    MentionKey:
      name: mentionKey
      in: path
      required: true
      schema:
        type: string
      description: Public mention key of an agent token
...
  securitySchemes:
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Standard human/agent login JWT
    BearerByoa:
      type: http
      scheme: bearer
      bearerFormat: BYOA
      description: Raw BYOA token with `byoa_` prefix
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    BatchMessageInput:
      type: object
      required: [roomId, content]
      properties:
        roomId:
          type: string
        content:
          type: string
        messageType:
          type: string
          enum: [TEXT, CODE, IMAGE, FILE, SYSTEM, AI_RESPONSE, RESEARCH_NOTE]
    BatchMessageRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/BatchMessageInput"
    BatchMessageResult:
      type: object
      required: [roomId]
      properties:
        roomId:
          type: string
        ok:
          type: boolean
        messageId:
          type: string
        error:
          type: string
    BatchMessageResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchMessageResult"
...
    AgentContextResponse:
      type: object
      required: [agent, rooms, tasks, memory]
      properties:
        agent:
          type: object
          required: [mentionKey, userId]
          properties:
            mentionKey:
              type: string
            userId:
              type: string
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/AgentContextRoom"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/AgentContextTask"
        memory:
          type: array
          items:
            $ref: "#/components/schemas/AgentContextMemoryEntry"
    AgentContextRoom:
      type: object
      required: [id, name, linkedProject, totalMessages, recentMessages]
      properties:
        id:
          type: string
        name:
          type: string
        linkedProject:
          type: object
          required: [id, name]
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
        totalMessages:
          type: integer
        recentMessages:
          type: array
          items:
            type: object
            required: [id, sender, senderType, content, timestamp]
            properties:
              id:
                type: string
              sender:
                type: string
              senderType:
                type: string
              content:
                type: string
              timestamp:
                type: string
                format: date-time
    AgentContextTask:
      type: object
      required: [id, title, status, priority, project]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        priority:
          type: string
        project:
          type: object
          required: [id, name]
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
    AgentContextMemoryEntry:
      type: object
      required:
        - id
        - scope
        - projectId
        - memoryType
        - title
        - tags
        - summary
        - confidence
        - pluginId
        - moduleKey
        - freshnessStatus
        - updatedAt
      properties:
        id:
          type: string
        scope:
          type: string
          enum: [GLOBAL, PROJECT]
        projectId:
          type: string
          nullable: true
        memoryType:
          type: string
        title:
          type: string
        tags:
          type: array
          items:
            type: string
        summary:
          type: string
        confidence:
          type: number
        pluginId:
          type: string
        moduleKey:
          type: string
          nullable: true
        freshnessStatus:
          type: string
          enum: [fresh, stale, unknown]
        freshnessWarning:
          type: string
          nullable: true
        validUntil:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
...